init: made initial setup
